Reducer: be more aggressive in eliminating return statements
This change allows the reducer to eliminate non-void return statements
that are inside dead code blocks.
