+++ reducer/src/main/java/com/graphicsfuzz/reducer/reductionopportunities/Compatibility.java
@@ -91,6 +91,18 @@ public final class Compatibility {
       return false;
     }
 
+    if (matchesEitherDirection(first, second,
+        StmtReductionOpportunity.class::equals,
+        UnswitchifyReductionOpportunity.class::equals)) {
+      return false;
+    }
+
+    if (matchesEitherDirection(first, second,
+        StmtReductionOpportunity.class::equals,
+        OutlinedStatementReductionOpportunity.class::equals)) {
+      return false;
+    }
+
     return true;
   }
