+++ reducer/src/test/java/com/graphicsfuzz/reducer/reductionopportunities/StmtReductionOpportunitiesTest.java
@@ -494,5 +494,4 @@ public class StmtReductionOpportunitiesTest {
 
   }
 
-
 }
